login added to navbar :fire: :fire: :fire:
