+++ client/src/components/App.js
@@ -1,9 +1,10 @@
 import React from 'react';
-import './App.css';
+import '../css/App.css';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import Home from './Home';
 import AppNavbar from './AppNavbar';
 import ResourceDetailed from './ResourceDetailed';
+import Login from './Login';
 
 export default class App extends React.Component {
   render() {
@@ -13,6 +14,7 @@ export default class App extends React.Component {
           <Route path="/" exact component={Home} />
           <Route path="/navbar" exact component={AppNavbar} />
           <Route path="/resource" exact component={ResourceDetailed} />
+          <Route path="/login" exact component={Login} />
         </div>
       </Router>
     );
