+++ client/src/components/Home.js
@@ -1,0 +1,34 @@
+// @flow
+
+import React, { Component } from 'react';
+import '../css/Home.css';
+import AppNavbar from './AppNavbar';
+
+type Props = {};
+
+type State = {| apiResponse: string |};
+
+export default class Home extends Component<Props, State> {
+  // constructor(props: Props) {
+  //   super(props);
+  //   this.state = { apiResponse: '' };
+  // }
+
+  // UNSAFE_componentWillMount() {
+  //   this.callAPI();
+  // }  
+
+  // callAPI() {
+  //   fetch('http://localhost:9000/testAPI')
+  //     .then(res => res.text())
+  //     .then(res => this.setState({ apiResponse: res }));
+  // }
+
+  render() {
+    return (
+      <div className="App">
+        <AppNavbar />
+      </div>
+    );
+  }
+}
