+++ client/src/css/ResourceDetailed.css
